junit testing for OperatorList class
Testing is an essential part of any software.
Hence, I have begun integrating junit testing to the application. It uses
mavenCentral repo for checking out all the dependencies.

And the benefits of this integration is already seen. I have removed a couple of
errors from the OperatorList class which I had overlooked earlier.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>