Use short for numQues
Number of question would not exceed 100 for a session (generally).
Byte would be more appropriate primitive for numQues; But for the sake of
expansion later on, it would be reasonable to use short.

Also this differentiation helps in setting constructors later on. If numQues and
maxRange would both be int, setting them in the constructor would be difficult.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>